Made multiple selection in two ways, now when collection control is getting selected items variable it adds each item in selected items collection. So when user switches between view options selected items are saved

+++ FileExplorer/UI/Behaviors/DataGrid/MultipleSelectionDataGridBehavior.cs
@@ -1,17 +1,13 @@
-﻿using Helpers.General;
+﻿using FileExplorer.UI.Behaviors.BaseBehaviors;
+using Helpers.General;
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.Xaml.Interactivity;
 using System.Collections;
 
 namespace FileExplorer.UI.Behaviors.DataGrid
 {
-    public sealed class MultipleSelectionDataGridBehavior : Behavior<CommunityToolkit.WinUI.UI.Controls.DataGrid>
+    public sealed class
+        MultipleSelectionDataGridBehavior : MultipleSelectionBase<CommunityToolkit.WinUI.UI.Controls.DataGrid>
     {
-        /// <summary>
-        /// Selected items property that used to bind list item from ViewModel
-        /// </summary>
-        public IList SelectedItems { get; set; }
-
         protected override void OnAttached()
         {
             base.OnAttached();
@@ -24,15 +20,22 @@ namespace FileExplorer.UI.Behaviors.DataGrid
             AssociatedObject.SelectionChanged -= OnSelectionChanged;
         }
 
-        /// <summary>
-        /// Method that takes care of removing and adding new selected items from <see cref="GridView"/>
-        /// </summary>
-        /// <param name="sender"> Selection event sender (GridView) </param>
-        /// <param name="e"> Event argument that stores all information about selection </param>
-        private void OnSelectionChanged(object sender, SelectionChangedEventArgs e)
+        protected override void SelectItemInCollection(IList items)
         {
-            SelectedItems.RemoveRange(e.RemovedItems);
-            SelectedItems.AddRange(e.AddedItems);
+            AssociatedObject.SelectedItems.AddRange(items);
+        }
+
+        protected override void OnSelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if (!bindingSelectedItems)
+            {
+                SelectedItems.RemoveRange(e.RemovedItems);
+                SelectedItems.AddRange(e.AddedItems);
+            }
+            else
+            {
+                bindingSelectedItems = false;
+            }
         }
     }
 }
